correction bug temps

+++ trunk/babel-2014-minett_a/trunk/lib/time.h
@@ -7,8 +7,8 @@ class ITime
 public:
     virtual void setToMsTimeOfDay() = 0;
     virtual ITime & operator- (ITime const &) const = 0;
-    virtual unsigned int getMs() const = 0;
-    virtual void setMs(unsigned int ms) = 0;
+    virtual unsigned long long getMs() const = 0;
+    virtual void setMs(unsigned long long ms) = 0;
 };
 
 #ifdef OS_WINDOWS
